Complete overhaul
Okay, this is a fun story. I was looking for hours and days, why the
fuck my faction isn't showing anymore and if it was showing why the
weapons do not work. So I thought: "Fuck it, I rewrite everything in
seperate files. Maybe that helps"....it didn't.What did help in the end
though, was to get rid of the AddonPrefix in Addon Builder -,-