Updated 3den Preview and Random Gear Script
I have put in some new 3den Previews (needsome more. Client (UN) has no gear and Contractor needs a new pic).

I've also finally added a new Random Gear script on the Contractor that only applies in the editor (or when the unit is spawned in general...so it should also work on Zeus and Alive)